Move files static folder
