+++ cli/web/static/index.html
