+++ cli/web/static/script.js
