+++ cli/web/static/style.css
