+++ cli/web/static/tailwind.config.js
